feat(inc-update): handle translations in incremental update process

+++ test/e2e.utils.js
@@ -4,6 +4,9 @@ const shell = require('shelljs');
 const e2eEnv = require('./e2e.env');
 const wsApi = require('supertest')(e2eEnv.wsUrl);
 
+const START_WAFFLE_SERVER = (process.env.START_WAFFLE_SERVER === 'false') ? false : true;
+const DROP_MONGO_DATABASE = (process.env.DROP_MONGO_DATABASE === 'false') ? false : true;
+
 module.exports = {
   dropMongoDb,
   stopWaffleServer,
@@ -23,15 +26,21 @@ function sendDdfqlRequest(ddfql, onResponseReceived) {
 
 function startWaffleServer() {
   setUpEnvironmentVariables();
-  shell.exec(`./node_modules/.bin/forever start -t -o /dev/null -l /dev/null -a --uid "${e2eEnv.wsUid}" server.js`);
+  if (START_WAFFLE_SERVER) {
+    shell.exec(`./node_modules/.bin/forever start -t -o /dev/null -l /dev/null -a --uid "${e2eEnv.wsUid}" server.js`);
+  }
 }
 
 function stopWaffleServer() {
-  shell.exec(`./node_modules/.bin/forever stop "${e2eEnv.wsUid}"`);
+  if (START_WAFFLE_SERVER) {
+    shell.exec(`./node_modules/.bin/forever stop "${e2eEnv.wsUid}"`);
+  }
 }
 
 function dropMongoDb() {
-  shell.exec(`mongo ${e2eEnv.mongodb} --eval "db.dropDatabase()"`);
+  if (DROP_MONGO_DATABASE) {
+    shell.exec(`mongo ${e2eEnv.mongodb} --eval "db.dropDatabase()"`);
+  }
 }
 
 function setUpEnvironmentVariables() {
